requirements.txt
aniso8601==9.0.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==3.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
PyJWT==1.7.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.9
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
